creates ProjectPicker component
