+++ src/components/ProjectsView/ProjectsLayout/index.tsx
@@ -15,6 +15,7 @@ import VectorLogoAndText from "../VectorLogoAndText";
 import { ProjectPreview } from "../ProjectPreview";
 import { GridElement, MotionGridElement } from "../GridElement";
 import { ProjectDetailsLayout } from "../ProjectDetailsPreviewLayout";
+import ProjectPicker from "../ProjectPicker";
 
 // animation variants
 const pageVariants = {
@@ -35,57 +36,10 @@ const pageTransition = {
     duration: 0.5
 };
 
-const gridElementVariants = {
-    initial: {
-        x: '-50px',
-        y: '-50px',
-        opacity: 0,
-    },
-    in: {
-        x: 0,
-        y: 0,
-        opacity: 1,
-    }
-};
 
-const gridTransition = {
-    type: "spring",
-    damping: 20,
-    stiffness: 100
-};
-
-const listVariants = {
-    initial: {
-        scale: 0,
-        y: '200px'
-    },
-    in: {
-        transition: {
-            staggerChildren: 0.1, // this will animate each child with a delay of 0.1s
-            delayChildren: 0.3 // this will delay the animation of all children by 0.3s
-        },
-        y: 0,
-        scale: 1
-    }
-};
 interface ProjectsLayoutProps {
     projects: ProjectData;
 }
-
-
-const StyledFlex = createStyledMotionComponent(motion.div)(props => `
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: start;
-    padding-top: 1rem;
-    padding-bottom: 1rem;
-    box-sizing: border-box;
-    border-radius: 60px;
-    margin: 30rem;
-    background-color: ${adjustTransparency(props.theme.cardBackground, 0.9)};
-    z-index: 9999;
-`);
 
 
 const StyledContainer = createStyledMotionComponent(motion.div)(props => `
@@ -94,26 +48,11 @@ const StyledContainer = createStyledMotionComponent(motion.div)(props => `
     gap: 1rem;
 `);
 
-const GridElementContainer = createStyledMotionComponent(motion.div)(props => `
-    display: flex;
-    padding: 2rem;
-
-`)
-
-const MemoizedGridElement = memo(GridElement);
 
 export default function ProjectsLayout({ projects }: ProjectsLayoutProps) {
-    const theme = useTheme();
     const [selectedProject, setSelectedProject] = useState<ProjectType | null>(null);
     const cursorType = useCursorEffect();
 
-    const handleMouseEnter = (project: ProjectType) => {
-        setSelectedProject(project);
-    }
-
-    const handleMouseLeave = () => {
-        setSelectedProject(null);
-    }
 
     return (
         <CursorContext.Provider value={cursorType}>
@@ -124,23 +63,12 @@ export default function ProjectsLayout({ projects }: ProjectsLayoutProps) {
                 variants={pageVariants}
                 transition={pageTransition}
             >
-                <StyledFlex variants={listVariants}>
-                    {Object.values(projects).map((project, index) => (
-                        <GridElementContainer
-                            key={project.id}
-                            transition={gridTransition}
-                            custom={index}
-                            variants={gridElementVariants}
-                        >
-                            <MemoizedGridElement
-                                project={project}
-                                handleMouseEnter={handleMouseEnter}
-                                isActive={selectedProject?.id === project.id}
-                                handleMouseLeave={handleMouseLeave}
-                            />
-                        </GridElementContainer>
-                    ))}
-                </StyledFlex>
+                <ProjectPicker
+                    selectedProject={selectedProject}
+                    setSelectedProject={setSelectedProject}
+                    projects={projects}
+                />
+
                 {selectedProject && (
                     <ProjectDetailsLayout
                         project={selectedProject}
